remove a level of abstraction by building the sql statements within the router callbacks
my thinking here is if i ever want to test this, i don't want to test sql statements in isolation. i'd rather go to the trouble to stand up a bare test database and test server and then create tests that make network calls and check the state of the database. these statements and expected states are simple enough that i'm of the mind to just skip of a level of testing (if i ever even get around to it).